refactor: グラフを描画する関数をgraphsディレクトリに移動
